fix: reduce global quality debt
